+++ game/main.lua
@@ -6,11 +6,21 @@ function love.load()
     Menu = require "src/menu"
     Scoreboard = require "src/scoreboard"
 
-    BLACK = 2
-    WHITE = 1
     WINDOW_WIDTH = love.graphics.getWidth()
     WINDOW_HEIGHT = love.graphics.getHeight()
     STATE = "menu"
+
+    -- colour constants
+    BLACK = 2
+    WHITE = 1
+    -- enemy type constants
+    TARGETER = 1
+    SPREADER = 2
+    -- path type constants
+    DROP = 1
+    CURVED = 2
+    STRAFE_LEFT = 3
+    STRAFE_RIGHT = 4
 
     Game:load()
     Images:load()